Add TownController.java and other fix

+++ src/main/java/com/slavapleshkov/springboot/spring_rest/exception/ExceptionMessageHelper.java
@@ -4,7 +4,7 @@ public class ExceptionMessageHelper {
 
     private ExceptionMessageHelper() {}
 
-    public static String entityNotFoundMsg(Long id) {
+    public static String entityNotFoundMsg(int id) {
         return String.format("Entity with id: %d not found", id);
     }
